+++ dgdg-api/src/main/java/com/depromeet/dgdg/controller/LocalTestController.kt
@@ -3,8 +3,8 @@ package com.depromeet.dgdg.controller
 import com.depromeet.dgdg.controller.dto.response.BaseResponse
 import com.depromeet.dgdg.config.auth.RequiredAuth
 import com.depromeet.dgdg.config.auth.UserId
-import com.depromeet.dgdg.service.token.AuthTokenProvider
-import com.depromeet.dgdg.service.token.dto.AuthTokenPayload
+import com.depromeet.dgdg.provider.token.AuthTokenProvider
+import com.depromeet.dgdg.provider.token.dto.AuthTokenPayload
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RestController
