+++ dgdg-api/src/main/java/com/depromeet/dgdg/provider/token/AuthTokenProvider.kt
@@ -1,4 +1,4 @@
-package com.depromeet.dgdg.service.token
+package com.depromeet.dgdg.provider.token
 
 interface AuthTokenProvider<T> {
